+++ include/qinterface.hpp
@@ -2456,7 +2456,7 @@ public:
     /**
      *  Get the device index. ("-1" is default).
      */
-    virtual int GetDeviceID() { return -1; }
+    virtual int64_t GetDeviceID() { return -1; }
 
     /**
      *  Get maximum number of amplitudes that can be allocated on current device.
